fix(composeExercise.js): fix passed questions not able to be answered again
It is possible to input a character from other questions.

However, it was not possible to input a character from any passed question,
because these characters were removed from the input table.

This caused an issue for the last question,
since the "next" question of the last question is currently itself.

This issue is fixed by separating the input table and the question list.

* MapCharacterTable()
    * add this.questList (init-ed to have the same content as this.table)
* MapCharacterTable.prototype.getCharacterAndAlphabet()
    * make the previous question deleted from this.questList instead